Change case for all files to kebab-case
