+++ src/helpers/is-emoji.ts
