+++ src/helpers/save-original-favicon.ts
