+++ src/reset-favicon.ts
